+++ frontend/src/components/modules/characters/panelSections/global/GlobalSection.tsx
@@ -1,6 +1,6 @@
 "use client"
 
-import Champs from "@/components/common/Champs";
+import Champs from "@/components/common/Field";
 import IClassification from "@/models/characters/classification/IClassification";
 import { useTranslations } from "next-intl";
 import ClassificationCard from "./cards/ClassificationCard";
